refactor(content): extract hideGnMetadata and colorizeThread from index.ts
Extract hideGnMetadataInComment into metadata-hider.ts with hoisted regex.
Extract colorizeCommentThread and clearCommentColorIndicators into
thread-colorizer.ts. Update index.ts to import from the new modules.
