+++ packages/browser-extension/src/content/index.ts
@@ -2,7 +2,9 @@ import { detectGitHubPage, type GitHubPageInfo } from './detector';
 import { observeDiffContent } from './diff-observer';
 import { getSelectionInfo } from './selection';
 import { scanForGnComments } from './comment-scanner';
-import { highlightTextRange, clearAllHighlights, HIGHLIGHT_COLORS } from './highlighter';
+import { highlightTextRange, clearAllHighlights } from './highlighter';
+import { hideGnMetadataInComment } from './metadata-hider';
+import { colorizeCommentThread, clearCommentColorIndicators } from './thread-colorizer';
 import { showOptInBanner, hideOptInBanner } from './ui/optin-banner';
 import { isRepoEnabled, enableRepo, isRepoBlocked, blockRepo } from '../storage/repo-settings';
 import { getHighlightStyle, applyHighlightStyle } from '../storage/highlight-style';
@@ -252,83 +254,5 @@ function scanAndHighlight(): void {
   }
 }
 
-function hideGnMetadataInComment(commentEl: HTMLElement): void {
-  // Don't hide metadata inside textareas — those are pending comments
-  // that need the ^gn tag visible in the editable text.
-  if (commentEl.closest('textarea') || commentEl.querySelector('textarea')) {
-    return;
-  }
-
-  const text = commentEl.textContent ?? '';
-  const cleaned = text.replace(/\^gn:\d+:[LR]:\d+:\d+/, '').trim();
-
-  if (!cleaned) {
-    commentEl.style.display = 'none';
-  } else {
-    const walker = document.createTreeWalker(commentEl, NodeFilter.SHOW_TEXT);
-    let node: Text | null;
-    while ((node = walker.nextNode() as Text | null)) {
-      if (node.textContent && /\^gn:\d+:[LR]:\d+:\d+/.test(node.textContent)) {
-        // Don't touch text nodes inside textareas or form inputs
-        if (node.parentElement?.closest('textarea, input')) continue;
-        const span = document.createElement('span');
-        span.style.display = 'none';
-        span.textContent = node.textContent;
-        node.replaceWith(span);
-        break;
-      }
-    }
-  }
-}
-
-/**
- * Apply a colored left border to the comment thread container to
- * visually associate it with its highlight color.
- *
- * For submitted comments: colors the author name link.
- * For pending comment boxes: colors the "Add a comment on line" heading.
- */
-function colorizeCommentThread(commentEl: HTMLElement, colorIndex: number): void {
-  const color = HIGHLIGHT_COLORS[colorIndex % HIGHLIGHT_COLORS.length];
-
-  // Walk up from the ^gn container to find the thread root
-  const threadContainer =
-    commentEl.closest('[data-marker-id]') ??
-    commentEl.closest('[data-testid="review-thread"]')?.parentElement ??
-    commentEl.closest('.inline-comment-form');
-  if (!threadContainer) return;
-
-  // Apply a colored left border to the thread container
-  (threadContainer as HTMLElement).style.borderLeft = `3px solid ${color}`;
-  (threadContainer as HTMLElement).setAttribute('data-gn-color-indicator', 'true');
-
-  // For submitted comments: color the author name
-  const authorLink = threadContainer.querySelector<HTMLElement>(
-    '[data-testid="avatar-link"], .ActivityHeader-module__AuthorName__VJr9h',
-  );
-  if (authorLink) {
-    authorLink.style.color = color;
-    authorLink.setAttribute('data-gn-color-indicator', 'true');
-  }
-
-  // For pending comment boxes: color the "Add a comment on line" heading
-  const heading = threadContainer.querySelector<HTMLElement>(
-    '.InlineReviewThread-module__inlineReviewThreadHeading__o7jqD, h4.prc-Heading-Heading-MtWFE',
-  );
-  if (heading) {
-    heading.style.color = color;
-    heading.setAttribute('data-gn-color-indicator', 'true');
-  }
-}
-
-/**
- * Remove all color indicators applied by colorizeCommentThread.
- */
-function clearCommentColorIndicators(): void {
-  const indicators = document.querySelectorAll<HTMLElement>('[data-gn-color-indicator]');
-  for (const el of indicators) {
-    el.style.borderLeft = '';
-    el.style.color = '';
-    el.removeAttribute('data-gn-color-indicator');
-  }
-}
+// hideGnMetadataInComment, colorizeCommentThread, and clearCommentColorIndicators
+// are now imported from ./metadata-hider and ./thread-colorizer
